test(EuiDataGrid): move EuiDataGrid render sorting arrows jest tests to a separate file
